+++ posthog-samples/posthog-android-sample/src/main/java/com/posthog/android/sample/MyCustomException.kt
@@ -1,0 +1,4 @@
+package com.posthog.android.sample
+
+class MyCustomException {
+}