+++ posthog-android/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+- feat: add manual captureException ([#300](https://github.com/PostHog/posthog-android/issues/300))
+
 ## 3.23.0 - 2025-10-06
 
 - feat: Android surveys use the new response question id format ([#289](https://github.com/PostHog/posthog-android/pull/289))